Refactor: Modularize visualization utils
